middle refactoring
