+++ pom.xml
@@ -9,10 +9,10 @@
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>pro.sky</groupId>
-    <artifactId>ExceptionsEmployeeHW5</artifactId>
+    <artifactId>MapEmployeeHW8</artifactId>
     <version>0.0.1-SNAPSHOT</version>
-    <name>ExceptionsEmployeeHW5</name>
-    <description>ExceptionsEmployeeHW5</description>
+    <name>MapEmployeeHW8</name>
+    <description>MapEmployeeHW8</description>
     <properties>
         <java.version>11</java.version>
     </properties>
