+++ src/main/java/pro/sky/MapEmployeeHW8/Employee.java
@@ -1,4 +1,4 @@
-package pro.sky.exceptionsemployeehw5;
+package pro.sky.MapEmployeeHW8;
 
 
 import java.util.Objects;
