+++ src/main/java/pro/sky/MapEmployeeHW8/EmployeeInterface.java
@@ -1,4 +1,6 @@
-package pro.sky.exceptionsemployeehw5;
+package pro.sky.MapEmployeeHW8;
+
+import java.util.Map;
 
 public interface EmployeeInterface {
     Employee addNewEmployee(String firstName, String lastName);
@@ -6,5 +8,8 @@ public interface EmployeeInterface {
     Employee findEmployee(String firstName, String lastName);
 
     Employee deleteEmployee(String firstName, String lastName);
+
+
+    Map<Employee, Integer> printAllEmployees();
 
 }
