+++ src/main/java/pro/sky/MapEmployeeHW8/ExceptionsEmployeeHW8Application.java
@@ -1,13 +1,13 @@
-package pro.sky.exceptionsemployeehw5;
+package pro.sky.MapEmployeeHW8;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class ExceptionsEmployeeHw5Application {
+public class ExceptionsEmployeeHW8Application {
 
     public static void main(String[] args) {
-        SpringApplication.run(ExceptionsEmployeeHw5Application.class, args);
+        SpringApplication.run(ExceptionsEmployeeHW8Application.class, args);
     }
 
 }
