+++ src/main/java/pro/sky/MapEmployeeHW8/exceptions/AlreadyExistsException.java
@@ -1,4 +1,4 @@
-package pro.sky.exceptionsemployeehw5.exceptions;
+package pro.sky.MapEmployeeHW8.exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
